Escape regex metacharacters before search.
Fixes #171
