+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "dependencies": {
     "d3": "^3.5.2",
+    "escape-string-regexp": "^1.0.3",
     "lesshat": "^3.0.2"
   },
   "devDependencies": {
